+++ src/types/engines.ts
@@ -1,13 +1,22 @@
-import { PlayerOptions, TrackStream } from "./player";
+import { StreamType } from "@discordjs/voice";
+import { Readable } from "stream";
+import { PlayerOptions } from "./player";
 
 export interface PlayerEngine {
+  /** Source (e.g. "youtube", "spotify" etc.) */
   source: string;
+  /**
+   * Whether this engine is responsible for searching/streaming the given query.
+   * If not available engines are responsible for a user query, YouTube will be automatically selected.
+   */
   isResponsible(query: string, playerOptions: PlayerOptions): Promise<boolean>;
+  /** Gets information about the given query. */
   search(
     query: string,
     playerOptions: PlayerOptions,
     options?: SearchOptions
   ): Promise<SearchResult[]>;
+  /** Gets the playable stream for the given track. */
   getStream(
     track: Track,
     playerOptions: PlayerOptions
@@ -19,7 +28,7 @@ export interface SearchOptions {
    * The source where tracks should be searched. If not provided, will automatically detect the source or fall back to YouTube.
    */
   source?: string;
-  /** Limit number of tracks to search. Does not work for Spotify. */
+  /** Limit number of tracks to search. Might not work for all engines. */
   limit?: number;
 }
 
@@ -52,4 +61,13 @@ export interface Playlist {
   title: string;
   url: string;
   thumbnailUrl?: string;
+}
+
+export interface TrackStream {
+  /**
+   * If the input is given as a string, then the inputType option will be overridden and FFmpeg will be used.
+   * Can be used as file path when playing local files.
+   */
+  stream: Readable | string;
+  type?: StreamType;
 }